+++ ego/src/solver/egor_impl.rs
